Completes implementation of all of the built-in datasets.
