+++ ludwig/datasets/configs/imbalanced_insurance.yaml
@@ -5,7 +5,7 @@ archive_filenames: imbalanced-data-practice.zip
 sha256:
   imbalanced-data-practice.zip: 33c7d15cbdb7cc151c1d5e920a8a613b015c19222f90d4eac04ca8cfc5416847
 dataset_filenames: aug_train.csv
-loader: imbalanced_insurance.ImbalancedInsuranceLoader
+loader: split_loaders.RandomSplitLoader
 description: |
   Health Insurance Cross Sell Prediction
   Predict Health Insurance Owners' who will be interested in Vehicle Insurance
