+++ ludwig/datasets/configs/santander_value_prediction.yaml
@@ -5,6 +5,7 @@ archive_filenames: santander-value-prediction-challenge.zip
 sha256:
   santander-value-prediction-challenge.zip: a8b44a0403bff6ab42f2bd1da8d9cbaf98f1fd4b9ea7a86e47491ac996384bf4
 train_filenames: train.csv
+loader: santander_value_prediction.SantanderValuePredictionLoader
 description: |
   The Santander Value Prediction Challenge dataset.
   https://www.kaggle.com/c/santander-value-prediction-challenge
