+++ ludwig/datasets/configs/sarcos.yaml
@@ -1,0 +1,18 @@
+version: 1.0
+name: sarcos
+download_urls:
+  - http://www.gaussianprocess.org/gpml/data/sarcos_inv.mat
+  - http://www.gaussianprocess.org/gpml/data/sarcos_inv_test.mat
+sha256:
+  sarcos_inv_test.mat: 161a59b5c3b4f4b404584323f181607b2acbe620eb134dc720760dc3f38f5cec
+  sarcos_inv.mat: b8a249733253ba6097372fedee7696833fcf30de42037d5b4a7227f21a6d1d97
+train_filenames: sarcos_inv.mat
+test_filenames: sarcos_inv_test.mat
+loader: sarcos.SarcosLoader
+description: |
+  The data relates to an inverse dynamics problem for a seven
+  degrees-of-freedom SARCOS anthropomorphic robot arm.
+  The task is to map from a 21-dimensional input space
+  (7 joint positions, 7 joint velocities, 7 joint accelerations)
+  to the corresponding 7 joint torques.
+  http://gaussianprocess.org/gpml/data/
