+++ ludwig/datasets/configs/sst2.yaml
@@ -1,0 +1,24 @@
+version: 1.0
+name: sst2
+download_urls: https://nlp.stanford.edu/~socherr/stanfordSentimentTreebank.zip
+sha256:
+  stanfordSentimentTreebank.zip: 3f5209483b46bbf129cacbbbe6ae02fe780407034f61cf6342b7833257c3f1db
+train_filenames: train.csv
+validation_filenames: dev.csv
+test_filenames: test.csv
+loader: sst.SST2Loader
+description: |
+  The SST2 dataset.
+
+  This dataset is constructed using the Stanford Sentiment Treebank Dataset.
+  This dataset contains binary labels (positive or negative) for each sample.
+
+  The original dataset specified 5 labels:
+  very negative, negative, neutral, positive, very positive with
+  the following cutoffs:
+  [0, 0.2], (0.2, 0.4], (0.4, 0.6], (0.6, 0.8], (0.8, 1.0]
+
+  In the construction of this dataset, we remove all neutral phrases
+  and assign a negative label if the original rating falls
+  into the following range: [0, 0.4] and a positive label
+  if the original rating is between (0.6, 1.0].
