+++ ludwig/datasets/configs/sst5.yaml
@@ -1,0 +1,19 @@
+version: 1.0
+name: sst5
+download_urls: https://nlp.stanford.edu/~socherr/stanfordSentimentTreebank.zip
+sha256:
+  stanfordSentimentTreebank.zip: 3f5209483b46bbf129cacbbbe6ae02fe780407034f61cf6342b7833257c3f1db
+train_filenames: train.csv
+validation_filenames: dev.csv
+test_filenames: test.csv
+loader: sst.SST5Loader
+description: |
+  The SST5 dataset.
+
+  This dataset is constructed using the Stanford Sentiment Treebank Dataset.
+  This dataset contains five labels (very negative, negative, neutral,
+  positive, very positive) for each sample.
+
+  In the original dataset, the  5 labels: very negative, negative, neutral, positive,
+  and very positive have the following cutoffs:
+  [0, 0.2], (0.2, 0.4], (0.4, 0.6], (0.6, 0.8], (0.8, 1.0]
