--- ludwig/datasets/noshow_appointments/config.yaml
@@ -1,4 +1,0 @@
-version: 1.0
-competition: joniarroba/noshowappointments
-archive_filename: noshowappointments.zip
-csv_filename: KaggleV2-May-2016.csv
