--- ludwig/datasets/rossmann_store_sales/config.yaml
@@ -1,4 +1,0 @@
-version: 1.0
-competition: rossmann-store-sales
-archive_filename: rossmann-store-sales.zip
-csv_filename: rossmann_store_sales.csv
